requirement.txt
altair==5.4.1
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.0
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.0.1
fonttools==4.40.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.4
ipython==8.24.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.13.5
nest-asyncio==1.6.0
numpy==2.2.1
opencv-python==4.8.0.74
openpyxl==3.1.5
packaging==23.1
pandas==2.2.3
parso==0.8.4
Pillow==10.0.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
protobuf==5.28.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==306
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.1
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
xgboost==2.1.3